CoastCapitalHomelab/requirements.txt
flask==3.1.0
gunicorn==23.0.0
anthropic==0.49.0
mysql-connector-python==9.2.0
requests==2.32.3
paramiko==3.5.1
prometheus_client==0.21.0
